refactor: separate the clothing item sizes types from the types file
